requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
langchain-community>=0.0.20
PyPDF2>=3.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
